fixed data type

+++ database/migrations/2017_12_26_232340_create_vests_table.php
@@ -20,7 +20,7 @@ class CreateVestsTable extends Migration
 
             // Datos Chaleco
             $table->integer('fit_id')->nullable();
-            $table->integer('talla')->nullable();
+            $table->double('talla')->nullable();
 
             $table->string('notas',1000)->nullable();
